112427: Avoid deleting source dir on item import

+++ dspace-api/src/main/java/org/dspace/app/itemimport/ItemImport.java
@@ -230,8 +230,7 @@ public class ItemImport extends DSpaceRunnable<ItemImportScriptConfiguration> {
             }
         } finally {
             // clean work dir
-            if (zip) {
-                FileUtils.deleteDirectory(new File(sourcedir));
+            if (zip && workDir != null) {
                 FileUtils.deleteDirectory(workDir);
             }
